Move depends to its own module
Fixes #116.
