+++ src/dependencies.js
@@ -1,0 +1,42 @@
+'use strict'
+
+var fs = require('fs-extra')
+var path = require('path')
+var semver = require('semver')
+
+module.exports = {
+  /**
+   * Determine the dependencies for the `shell.moveItemToTrash` API, based on the
+   * Electron version in use.
+   */
+  getTrashDepends: function getTrashDepends (options, callback) {
+    fs.readFile(path.resolve(options.src, 'version'), (err, tag) => {
+      if (err) return callback(err)
+
+      // The content of the version file is the tag name, e.g. "v1.8.1"
+      var version = tag.toString().slice(1).trim()
+      if (semver.lt(version, '1.4.1')) {
+        return callback(null, 'gvfs-bin')
+      } else if (semver.lt(version, '1.7.2')) {
+        return callback(null, 'kde-cli-tools | kde-runtime | trash-cli | gvfs-bin')
+      } else {
+        return callback(null, 'kde-cli-tools | kde-runtime | trash-cli | libglib2.0-bin | gvfs-bin')
+      }
+    })
+  },
+
+  /**
+   * Determine the default dependencies for an Electron application.
+   */
+  getDepends: function getDepends (trashDependencies) {
+    return [
+      trashDependencies,
+      'libgconf2-4',
+      'libgtk2.0-0',
+      'libnotify4',
+      'libnss3',
+      'libxtst6',
+      'xdg-utils'
+    ]
+  }
+}
